+++ HPXMLtoOpenStudio/measure.rb
@@ -212,7 +212,7 @@ class OSModel
     add_ceiling_fans(runner, model, weather, spaces)
 
     # Hot Water
-    hw_plant_loop = add_hot_water_and_appliances(runner, model, weather, spaces)
+    add_hot_water_and_appliances(runner, model, weather, spaces)
 
     # Plug Loads & Fuel Loads & Lighting
     add_mels(runner, model, spaces)
@@ -232,7 +232,6 @@ class OSModel
 
     # Output
     add_unmet_hours_output(model, spaces)
-    add_unmet_wh_loads(runner, model, weather, hw_plant_loop)
     add_loads_output(model, spaces, add_component_loads)
     set_output_files(model)
     # Uncomment to debug EMS
@@ -1308,9 +1307,10 @@ class OSModel
     end
 
     # Hot water fixtures and appliances
-    HotWaterAndAppliances.apply(model, runner, @hpxml, weather, spaces, hot_water_distribution,
+    showers_peak_flow = nil #used for unmet wh load calculations
+    showers_peak_flow = HotWaterAndAppliances.apply(model, runner, @hpxml, weather, spaces, hot_water_distribution,
                                 solar_thermal_system, @eri_version, @schedules_file, plantloop_map,
-                                @hpxml.header.unavailable_periods)
+                                @hpxml.header.unavailable_periods) 
 
     if (not solar_thermal_system.nil?) && (not solar_thermal_system.collector_area.nil?) # Detailed solar water heater
       loc_space, loc_schedule = get_space_or_schedule_from_location(solar_thermal_system.water_heating_system.location, model, spaces)
@@ -1320,7 +1320,9 @@ class OSModel
     # Add combi-system EMS program with water use equipment information
     Waterheater.apply_combi_system_EMS(model, @hpxml.water_heating_systems, plantloop_map)
 
-    return plantloop_map[sys_id]
+    #Add unmet wh loads calculation
+    Waterheater.unmet_wh_loads_program(model, @hpxml.water_heating_systems, plantloop_map, showers_peak_flow)
+    
   end
 
   def self.add_cooling_system(model, spaces, airloop_map)
@@ -1891,105 +1893,6 @@ class OSModel
     program_calling_manager.setName("#{program.name} calling manager")
     program_calling_manager.setCallingPoint('EndOfZoneTimestepBeforeZoneReporting')
     program_calling_manager.addProgram(program)
-  end
-
-  def self.add_unmet_wh_loads(runner, model, weather, hw_plant_loop)
-    @hpxml.water_heating_systems.each do |water_heating_system|
-      #Get shower schedule max value
-      shower_peak_flow = HotWaterAndAppliances.add_showers_and_calculate_max(model, runner, @hpxml, weather, water_heating_system, @eri_version, @schedules_file, @hpxml.header.unavailable_periods)
-      sys_id = water_heating_system.id
-      #Get the water storage tanks for the outlet temp sensor
-      num_tanks = 0
-      tank = nil
-
-      hw_plant_loop.components.each do |c|
-        next unless c.to_WaterHeaterMixed.is_initialized
-        tank = c.to_WaterHeaterMixed.get
-        num_tanks += 1
-      end
-      hw_plant_loop.components.each do |c|
-        next unless c.to_WaterHeaterStratified.is_initialized
-        tank = c.to_WaterHeaterStratified.get
-        num_tanks += 1
-      end
-
-      if num_tanks > 1
-        runner.registerError("Found more WHs than expected. There are #{num_tanks} water heaters instead of 1.")
-      end
-
-      # EMS sensors
-      mixed_setpoint_sensor = OpenStudio::Model::EnergyManagementSystemSensor.new(model, 'Schedule Value')
-      mixed_setpoint_sensor.setName('res_shower_mixsp')
-      mixed_setpoint_sensor.setKeyName('mixed water temperature schedule')
-
-      wh_temp_sensor = OpenStudio::Model::EnergyManagementSystemSensor.new(model, 'Water Heater Use Side Outlet Temperature')
-      wh_temp_sensor.setName("#{tank.name} Outlet Temperature")
-      wh_temp_sensor.setKeyName("#{tank.name}")
-
-      shower_flow_sensor = OpenStudio::Model::EnergyManagementSystemSensor.new(model, 'Schedule Value')
-      shower_flow_sensor.setName('Shower Volume')
-      shower_flow_sensor.setKeyName('hot_water_showers')
-
-      #cw_flow_sensor = OpenStudio::Model::EnergyManagementSystemSensor.new(model, 'Water Use Equipment Hot Water Volume')
-      #cw_flow_sensor.setName('Clothes Washer Volume')
-      #cw_flow_sensor.setKeyName('clothes washer')
-
-      #dw_flow_sensor = OpenStudio::Model::EnergyManagementSystemSensor.new(model, 'Water Use Equipment Hot Water Volume')
-      #dw_flow_sensor.setName('Dishwasher Volume')
-      #dw_flow_sensor.setKeyName('dishwasher')
-
-      #fx_flow_sensor = OpenStudio::Model::EnergyManagementSystemSensor.new(model, 'Water Use Equipment Hot Water Volume')
-      #fx_flow_sensor.setName('Fixture Volume')
-      #fx_flow_sensor.setKeyName('dhw fixtures')
-
-      # EMS program
-      unmet_wh_loads_program = OpenStudio::Model::EnergyManagementSystemProgram.new(model)
-      unmet_wh_loads_program.setName(Constants.ObjectNameUnmetLoadsProgram)
-      unmet_wh_loads_program.addLine("If #{shower_flow_sensor.name} > 0")
-      unmet_wh_loads_program.addLine('Set ShowerTime=SystemTimeStep')
-      unmet_wh_loads_program.addLine('Else')
-      unmet_wh_loads_program.addLine('Set ShowerTime=0')
-      unmet_wh_loads_program.addLine('EndIf')
-      unmet_wh_loads_program.addLine("If (#{shower_flow_sensor.name} > 0) && (#{wh_temp_sensor.name} < #{mixed_setpoint_sensor.name})")
-      #unmet_wh_loads_program.addLine("If (#{shower_flow_sensor.name} > 0) && (49 < #{mixed_setpoint_sensor.name})")
-      unmet_wh_loads_program.addLine('Set ShowerSagTime=SystemTimeStep')
-      #unmet_wh_loads_program.addLine("Set ShowerE=#{shower_flow_sensor.name} * #{shower_peak_flow} * 4141170 * (49 - #{mixed_setpoint_sensor.name})")
-      unmet_wh_loads_program.addLine("Set ShowerE=#{shower_flow_sensor.name} * #{shower_peak_flow} * 4141170 * (#{wh_temp_sensor.name} - #{mixed_setpoint_sensor.name})")
-      unmet_wh_loads_program.addLine('Else')
-      unmet_wh_loads_program.addLine('Set ShowerSagTime=0')
-      unmet_wh_loads_program.addLine('Set ShowerE=0')
-      unmet_wh_loads_program.addLine('EndIf')
-
-      # ProgramCallingManagers
-      program_calling_manager = OpenStudio::Model::EnergyManagementSystemProgramCallingManager.new(model)
-      program_calling_manager.setName("#{unmet_wh_loads_program.name} calling manager")
-      program_calling_manager.setCallingPoint('EndOfSystemTimestepAfterHVACReporting')
-      program_calling_manager.addProgram(unmet_wh_loads_program)
-
-      # # output variables
-      # shower_time_output_var = OpenStudio::Model::EnergyManagementSystemOutputVariable.new(model, 'ShowerTime')
-      # shower_time_output_var.setName('ShowerTime')
-      # shower_time_output_var.setTypeOfDataInVariable('Summed')
-      # shower_time_output_var.setUpdateFrequency('SystemTimestep')
-      # shower_time_output_var.setEMSProgramOrSubroutineName(unmet_wh_loads_program)
-      # shower_time_output_var.setUnits('hr')
-
-      # # output variables
-      # shower_unmet_energy_output_var = OpenStudio::Model::EnergyManagementSystemOutputVariable.new(model, 'ShowerE')
-      # shower_unmet_energy_output_var.setName('ShowerE')
-      # shower_unmet_energy_output_var.setTypeOfDataInVariable('Summed')
-      # shower_unmet_energy_output_var.setUpdateFrequency('SystemTimestep')
-      # shower_unmet_energy_output_var.setEMSProgramOrSubroutineName(unmet_wh_loads_program)
-      # shower_unmet_energy_output_var.setUnits('J')
-
-      # # output variables
-      # shower_unmet_time_output_var = OpenStudio::Model::EnergyManagementSystemOutputVariable.new(model, 'ShowerSagTime')
-      # shower_unmet_time_output_var.setName('ShowerSagTime')
-      # shower_unmet_time_output_var.setTypeOfDataInVariable('Summed')
-      # shower_unmet_time_output_var.setUpdateFrequency('SystemTimestep')
-      # shower_unmet_time_output_var.setEMSProgramOrSubroutineName(unmet_wh_loads_program)
-      # shower_unmet_time_output_var.setUnits('hr')
-    end
   end
 
   def self.add_loads_output(model, spaces, add_component_loads)