+++ HPXMLtoOpenStudio/resources/hotwater_appliances.rb
@@ -280,27 +280,57 @@ class HotWaterAndAppliances
         runner.registerWarning("Both '#{fixtures_col_name}' schedule file and monthly multipliers provided; the latter will be ignored.") if !hpxml.water_heating.water_fixtures_monthly_multipliers.nil?
       end
 
+      #Create shower schedule, used only for unmet load calculations
+      # Create separate shower schedule: Only used for calculating unmet loads. Shower hot water usage is part of the fixtures usage.
+      showers_schedule = nil
+      showers_col_name = SchedulesFile::ColumnHotWaterShowers
+      if not schedules_file.nil?
+        showers_schedule = schedules_file.create_schedule_file(col_name: showers_col_name, schedule_type_limits_name: Constants.ScheduleTypeLimitsFraction)
+      end
+      if showers_schedule.nil?
+        showers_unavailable_periods = Schedule.get_unavailable_periods(runner, showers_col_name, unavailable_periods)
+        showers_weekday_sch = Schedule.ShowersWeekdayFractions
+        showers_weekend_sch = Schedule.ShowersWeekendFractions
+        showers_monthly_sch = Schedule.ShowersMonthlyMultipliers
+        showers_schedule_obj = MonthWeekdayWeekendSchedule.new(model, Constants.ObjectNameShowers, showers_weekday_sch, showers_weekend_sch, showers_monthly_sch, Constants.ScheduleTypeLimitsFraction, unavailable_periods: unavailable_periods)
+        showers_schedule = showers_schedule_obj.schedule
+      else
+        runner.registerWarning("Both '#{showers_col_name}' schedule file and weekday fractions provided; the latter will be ignored.") if !Schedule.ShowersWeekdayFractions.nil?
+        runner.registerWarning("Both '#{showers_col_name}' schedule file and weekend fractions provided; the latter will be ignored.") if !Schedule.ShowersWeekendFractions.nil?
+        runner.registerWarning("Both '#{showers_col_name}' schedule file and monthly multipliers provided; the latter will be ignored.") if !Schedule.ShowersMonthlyMultipliers.nil?
+      end
     end
 
+    #create an array of peak flow to return
+    shower_peak_flows = {}
+    num_wh = 0
     hpxml.water_heating_systems.each do |water_heating_system|
       non_solar_fraction = 1.0 - Waterheater.get_water_heater_solar_fraction(water_heating_system, solar_thermal_system)
 
       gpd_frac = water_heating_system.fraction_dhw_load_served # Fixtures fraction
       if gpd_frac > 0
 
+        #For showers, calculate flow rates but don't add a WaterUse:Equipment object. Shower usage is included in fixtures and only used for tracking unmet loads
         fx_gpd = get_fixtures_gpd(eri_version, nbeds, fixtures_all_low_flow, daily_mw_fractions, fixtures_usage_multiplier)
+        shower_gpd = get_showers_gpd(eri_version, nbeds, fixtures_all_low_flow, daily_mw_fractions, fixtures_usage_multiplier)
         w_gpd = get_dist_waste_gpd(eri_version, nbeds, has_uncond_bsmnt, cfa, ncfl, hot_water_distribution, fixtures_all_low_flow, fixtures_usage_multiplier)
 
         fx_peak_flow = nil
+        shower_peak_flow = nil
         if not schedules_file.nil?
           fx_peak_flow = schedules_file.calc_peak_flow_from_daily_gpm(col_name: SchedulesFile::ColumnHotWaterFixtures, daily_water: fx_gpd)
+          shower_peak_flow = schedules_file.calc_peak_flow_from_daily_gpm(col_name: SchedulesFile::ColumnHotWaterFixtures, daily_water: shower_gpd)
           dist_water_peak_flow = schedules_file.calc_peak_flow_from_daily_gpm(col_name: SchedulesFile::ColumnHotWaterFixtures, daily_water: w_gpd)
         end
         if fx_peak_flow.nil?
           fx_peak_flow = fixtures_schedule_obj.calc_design_level_from_daily_gpm(fx_gpd)
+          shower_peak_flow = showers_schedule_obj.calc_design_level_from_daily_gpm(shower_gpd)
           dist_water_peak_flow = fixtures_schedule_obj.calc_design_level_from_daily_gpm(w_gpd)
         end
 
+        id = water_heating_system.id
+        shower_peak_flows[:id] = shower_peak_flow
+        
         # Fixtures (showers, sinks, baths)
         add_water_use_equipment(model, Constants.ObjectNameFixtures, fx_peak_flow * gpd_frac * non_solar_fraction, fixtures_schedule, water_use_connections[water_heating_system.id], mw_temp_schedule)
 
@@ -321,6 +351,7 @@ class HotWaterAndAppliances
             dist_pump.additionalProperties.setFeature('HPXML_ID', water_heating_system.id) # Used by reporting measure
           end
         end
+        num_wh += 1
       end
 
       # Clothes washer
@@ -390,6 +421,8 @@ class HotWaterAndAppliances
       add_other_equipment(model, Constants.ObjectNameWaterSensible, conditioned_space, water_design_level_sens, 1.0, 0.0, fixtures_schedule, nil)
       add_other_equipment(model, Constants.ObjectNameWaterLatent, conditioned_space, water_design_level_lat, 0.0, 1.0, fixtures_schedule, nil)
     end
+
+    return shower_peak_flows
   end
 
   def self.get_range_oven_default_values()
@@ -965,69 +998,8 @@ class HotWaterAndAppliances
     return f_eff * ref_f_gpd * fixtures_usage_multiplier
   end
 
-  def self.add_showers_and_calculate_max(model, runner, hpxml, weather, water_heating_system, eri_version, schedules_file, unavailable_periods)
-    nbeds = hpxml.building_construction.additional_properties.adjusted_number_of_bedrooms
-
-    # Process fixtures
-    fixtures_usage_multiplier = hpxml.water_heating.water_fixtures_usage_multiplier
-    fixtures_all_low_flow = true
-    hpxml.water_fixtures.each do |water_fixture|
-      next unless [HPXML::WaterFixtureTypeShowerhead, HPXML::WaterFixtureTypeFaucet].include? water_fixture.water_fixture_type
-
-      fixtures_all_low_flow = false if not water_fixture.low_flow
-    end
-
-    # Get distribution system
-    if hpxml.water_heating_systems.size > 0
-      hot_water_distribution = hpxml.hot_water_distributions[0]
-    end
-
-    # Calculate mixed water fractions
-    t_mix = 105.0 # F, Temperature of mixed water at fixtures
-    avg_setpoint_temp = 0.0 # WH Setpoint: Weighted average by fraction DHW load served
-    hpxml.water_heating_systems.each do |water_heating_system|
-      wh_setpoint = water_heating_system.temperature
-      wh_setpoint = Waterheater.get_default_hot_water_temperature(eri_version) if wh_setpoint.nil? # using detailed schedules
-      avg_setpoint_temp += wh_setpoint * water_heating_system.fraction_dhw_load_served
-    end
-    daily_wh_inlet_temperatures = calc_water_heater_daily_inlet_temperatures(weather, nbeds, hot_water_distribution, fixtures_all_low_flow, hpxml.header.sim_calendar_year)
-    daily_mw_fractions = calc_mixed_water_daily_fractions(daily_wh_inlet_temperatures, avg_setpoint_temp, t_mix)
-
-    # Create separate shower schedule: Only used for calculating unmet loads. Shower hot water usage is part of the fixtures usage.
-    showers_schedule = nil
-    showers_col_name = SchedulesFile::ColumnHotWaterShowers
-    if not schedules_file.nil?
-      showers_schedule = schedules_file.create_schedule_file(col_name: showers_col_name, schedule_type_limits_name: Constants.ScheduleTypeLimitsFraction)
-    end
-    if showers_schedule.nil?
-      showers_unavailable_periods = Schedule.get_unavailable_periods(runner, showers_col_name, unavailable_periods)
-      showers_weekday_sch = Schedule.ShowersWeekdayFractions
-      showers_weekend_sch = Schedule.ShowersWeekendFractions
-      showers_monthly_sch = Schedule.ShowersMonthlyMultipliers
-      showers_schedule_obj = MonthWeekdayWeekendSchedule.new(model, Constants.ObjectNameShowers, showers_weekday_sch, showers_weekend_sch, showers_monthly_sch, Constants.ScheduleTypeLimitsFraction, unavailable_periods: unavailable_periods)
-      showers_schedule = showers_schedule_obj.schedule
-    else
-      runner.registerWarning("Both '#{showers_col_name}' schedule file and weekday fractions provided; the latter will be ignored.") if !Schedule.ShowersWeekdayFractions.nil?
-      runner.registerWarning("Both '#{showers_col_name}' schedule file and weekend fractions provided; the latter will be ignored.") if !Schedule.ShowersWeekendFractions.nil?
-      runner.registerWarning("Both '#{showers_col_name}' schedule file and monthly multipliers provided; the latter will be ignored.") if !Schedule.ShowersMonthlyMultipliers.nil?
-    end
-
-    gpd_frac = water_heating_system.fraction_dhw_load_served # Fixtures fraction
-    if gpd_frac > 0
-      shower_gpd = get_showers_gpd(eri_version, nbeds, fixtures_all_low_flow, daily_mw_fractions, fixtures_usage_multiplier)
-      shower_peak_flow = nil
-      if not schedules_file.nil?
-        shower_peak_flow = schedules_file.calc_peak_flow_from_daily_gpm(col_name: SchedulesFile::ColumnHotWaterFixtures, daily_water: shower_gpd)
-      end
-      if shower_peak_flow.nil?
-        shower_peak_flow = showers_schedule_obj.calc_design_level_from_daily_gpm(shower_gpd)
-      end
-    end
-
-    return shower_peak_flow
-  end
-
   def self.get_showers_gpd(eri_version, nbeds, fixtures_all_low_flow, daily_mw_fractions, fixtures_usage_multiplier = 1.0)
+    #JEFF
     if nbeds < 0.0
       return 0.0
     end